added atuhentication to delete and update todo endpoints

+++ bestseller_udemy/todo/TodoApp/routers/todos.py
@@ -61,10 +61,13 @@ async def create_todo(user: user_dependency, db:db_dependency, todo_request: Tod
 
 
 @router.put('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, 
+async def update_todo(user: user_dependency,
+                      db: db_dependency, 
                        todo_request: TodoRequest, # this line needs to be above path validation
                        todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='authentication failed')
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code=404, detail='No todo found')
     todo.title = todo_request.title
@@ -77,8 +80,10 @@ async def update_todo(db: db_dependency,
 
 
 @router.delete('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='authentication failed')
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code=404, detail='No todo found')
     db.delete(todo)
